tools/ocr_sidecar_compare/pyproject.toml
[project]
name = "ocr-sidecar-compare"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "paddlex[ocr-core]==3.3.13",
  "paddleocr==3.3.0",
]

[dependency-groups]
build = [
  "pyinstaller==6.16.0",
]
cpu = [
  "paddlepaddle==3.2.2",
]
gpu = [
  "paddlepaddle-gpu==3.2.2",
  "nvidia-cublas-cu11==11.11.3.6; sys_platform == 'win32'",
  "nvidia-cuda-nvrtc-cu11==11.8.89; sys_platform == 'win32'",
  "nvidia-cuda-runtime-cu11==11.8.89; sys_platform == 'win32'",
  "nvidia-cudnn-cu11==8.9.4.19; sys_platform == 'win32'",
  "nvidia-cufft-cu11==10.9.0.58; sys_platform == 'win32'",
  "nvidia-curand-cu11==10.3.0.86; sys_platform == 'win32'",
  "nvidia-cusolver-cu11==11.4.1.48; sys_platform == 'win32'",
  "nvidia-cusparse-cu11==11.7.5.86; sys_platform == 'win32'",
]

[tool.uv]
package = false
environments = [
  "python_full_version >= '3.11' and sys_platform == 'win32' and platform_machine == 'AMD64'",
  "python_full_version >= '3.11' and sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.uv.sources]
paddlepaddle-gpu = { index = "paddle-cu118" }
nvidia-cublas-cu11 = { index = "paddle-cu118" }
nvidia-cuda-nvrtc-cu11 = { index = "paddle-cu118" }
nvidia-cuda-runtime-cu11 = { index = "paddle-cu118" }
nvidia-cudnn-cu11 = { index = "paddle-cu118" }
nvidia-cufft-cu11 = { index = "paddle-cu118" }
nvidia-curand-cu11 = { index = "paddle-cu118" }
nvidia-cusolver-cu11 = { index = "paddle-cu118" }
nvidia-cusparse-cu11 = { index = "paddle-cu118" }

[[tool.uv.index]]
name = "paddle-cu118"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu118/"
explicit = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
